+++ src/views/admin/EditCreateAccountForm.vue
@@ -39,6 +39,8 @@
                 v-model:matriculation-id="account.student.matriculationId"
                 :edit-mode="editMode"
                 :error-bag="errorBag"
+                :username="account.user.username"
+                :latest="account.student.latestImmatriculation"
             />
             <section class="py-8 border-t-2 border-gray-400" :hidden="!editMode">
                 <div class="lg:flex">